Implemented file metadata capture functionality.
When the user POSTs a file to /api/fileanalyse, with the help of multer,
the app stores the file on the server, gets the file metadata from the
request object, responds to the user with a JSON object containing said
metadata, then unlinks the file from disk.

Comments were also rewritten to match my own style.
sample.env is now simply .env, which is not tracked.
.gitignore ignores .DS_Store files on Mac.